+++ src/main/java/emu/nes/addressing/Immediate.java
@@ -14,7 +14,7 @@ public class Immediate implements Addressing {
     public AddressingResult address(Bus bus, Registers registers, int pc) {
         registers.setPc(pc + 2);
         final int addr = pc + 1;
-        final int data = bus.read(addr);
+        final byte data = bus.read(addr);
         AddressingResult result = new AddressingResult();
         result.address = addr;
         result.data = data;
